Use integers as ip-version in os_subnet
Neutron expects integers as ip-version (4 or 6) in a subnet create[1]
but os_subnet module expects strings as ip-version ("4" or "6") in
playbooks and uses these strings instead of integers in subnet create
http requests. It's currently working because Neutron is a bit
permissive.

This commit uses integers instead of strings as ip-version in os_subnet.

[1] http://developer.openstack.org/api-ref-networking-v2.html#createSubnet